Fixes and minor layout improvements

+++ gtfs.py
@@ -72,7 +72,7 @@ def load(system, force_download=False, update_db=False):
         print(f'Failed to load GTFS for {system}: {e}')
 
 def download(system):
-    '''Downloads the GTFS for the given system, then loads it into memory'''
+    '''Downloads the GTFS for the given system'''
     if not system.gtfs_enabled:
         return
     data_zip_path = f'data/gtfs/{system.id}.zip'
@@ -99,6 +99,9 @@ def download(system):
         print(f'Failed to download GTFS for {system}: {e}')
 
 def update_database(system):
+    '''Updates cached GTFS data for the given system'''
+    if not system.gtfs_enabled:
+        return
     print(f'Updating database with GTFS data for {system}')
     try:
         helpers.departure.delete_all(system)
